python_at_demo/requirements.txt
# Function dependencies, for example:
# package>=version
Google-cloud-build
google-cloud-logging

